Split up Web3Config
