+++ archetypes/Pools/PoolsTable/index.tsx
@@ -1,8 +1,7 @@
 import React, { useState, useEffect, useMemo } from 'react';
-import { CommitActionEnum, SideEnum } from '@tracer-protocol/pools-js';
+import { CommitActionEnum, NETWORKS, SideEnum } from '@tracer-protocol/pools-js';
 import Button from '@components/General/Button';
 import { Table, TableHeader, TableRow, TableHeaderCell, TableRowCell } from '@components/General/TWTable';
-import { ARBITRUM } from '~/constants/networks';
 import { calcPercentageDifference, getPriceFeedUrl, toApproxCurrency } from '~/utils/converters';
 import { BrowseTableRowData, DeltaEnum } from '../state';
 import { TWModal } from '@components/General/TWModal';
@@ -26,6 +25,7 @@ import { StyledTooltip } from '@components/Tooltips';
 import { default as UpOrDown } from '@components/UpOrDown';
 import Info from '/public/img/general/info.svg';
 import LinkIcon from '@public/img/general/link.svg';
+import { networkConfig } from '@context/Web3Context/Web3Context.Config';
 
 type TProps = {
     onClickMintBurn: (pool: string, side: SideEnum, commitAction: CommitActionEnum) => void;
@@ -94,7 +94,7 @@ export default (({ rows, onClickMintBurn, showNextRebalance, deltaDenotation }) 
     const [showModalEffectiveGain, setShowModalEffectiveGain] = useState(false);
     const [showModalPoolDetails, setShowModalPoolDetails] = useState(false);
     const [poolDetails, setPoolDetails] = useState<any>({});
-    const { provider, account, config } = useWeb3();
+    const { provider, account, network = NETWORKS.ARBITRUM } = useWeb3();
 
     const handlePoolDetailsClick = (data: any) => {
         setShowModalPoolDetails(true);
@@ -266,7 +266,7 @@ export default (({ rows, onClickMintBurn, showNextRebalance, deltaDenotation }) 
                 open={showModalPoolDetails}
                 onClose={() => setShowModalPoolDetails(false)}
                 poolDetails={poolDetails}
-                previewUrl={config?.previewUrl || ''}
+                previewUrl={networkConfig[network]?.previewUrl || ''}
             />
         </>
     );
@@ -616,7 +616,7 @@ const TokenRows: React.FC<
                             <LinkOutlined
                                 className="align-middle ml-1"
                                 onClick={() => {
-                                    open(constructBalancerLink(tokenInfo.address, ARBITRUM, true), 'blank');
+                                    open(constructBalancerLink(tokenInfo.address, NETWORKS.ARBITRUM, true), 'blank');
                                 }}
                             />
                         </>
