+++ constants/index.ts
@@ -1,5 +1,5 @@
 export * from './commits';
-export * from './farms';
+export * from './staking';
 export * from './networks';
 export * from './general';
 export * from './pools';
