+++ context/ArbitrumBridgeContext/index.tsx
@@ -1,19 +1,23 @@
 import React, { useContext, useCallback, useState, useMemo } from 'react';
 import { ethers } from 'ethers';
 import BigNumber from 'bignumber.js';
-import { useWeb3 } from '../Web3Context/Web3Context';
-import { networkConfig, Network } from '../Web3Context/Web3Context.Config';
+import { Bridge, L1TokenData, L2TokenData, Inbox__factory } from 'arb-ts';
+import { NETWORKS } from '@tracer-protocol/pools-js';
 import { ERC20__factory } from '@tracer-protocol/perpetual-pools-contracts/types';
-import { destinationNetworkLookup, bridgeableAssets, bridgeableTickers } from '~/constants/bridge';
 
 import { isArbitrumNetwork } from '~/utils/supportedNetworks';
-import { Bridge, L1TokenData, L2TokenData, Inbox__factory } from 'arb-ts';
 import { Children } from '~/types/general';
 import { BridgeableAsset, BridgeableAssets, BridgeableBalances } from '~/types/bridge';
-import { ARBITRUM, MAINNET, MAX_SOL_UINT } from '~/constants/index';
+
+import { destinationNetworkLookup, bridgeableAssets, bridgeableTickers } from '~/constants/bridge';
+import { MAX_SOL_UINT } from '~/constants/general';
+
 import { useStore } from '@store/main';
 import { TransactionType } from '@store/TransactionSlice/types';
 import { selectHandleTransaction } from '@store/TransactionSlice';
+
+import { useWeb3 } from '../Web3Context/Web3Context';
+import { networkConfig, Network } from '../Web3Context/Web3Context.Config';
 
 type CachedBridges = {
     [account: string]: {
@@ -41,8 +45,8 @@ export const ArbitrumBridgeContext = React.createContext<ArbitrumBridgeProps>({
         console.debug(`arbitrumBridge.approve not ready`, tokenAddress, spender),
     refreshBridgeableBalance: async (asset: BridgeableAsset) =>
         console.debug(`arbitrumBridge.refreshBridgeableBalance not ready`, asset),
-    fromNetwork: networkConfig[MAINNET],
-    toNetwork: networkConfig[ARBITRUM],
+    fromNetwork: networkConfig[NETWORKS.MAINNET],
+    toNetwork: networkConfig[NETWORKS.ARBITRUM],
     bridgeableAssets: {},
     bridgeableBalances: {},
 });
@@ -56,7 +60,7 @@ const BRIDGEABLE_ASSET_ETH = {
 };
 
 export const ArbitrumBridgeStore: React.FC = ({ children }: Children) => {
-    const { account, signer, provider, network = MAINNET } = useWeb3();
+    const { account, signer, provider, network = NETWORKS.MAINNET } = useWeb3();
 
     const handleTransaction = useStore(selectHandleTransaction);
