+++ context/Web3Context/Web3Context.tsx
@@ -1,16 +1,11 @@
-// https://choosealicense.com/licenses/lgpl-3.0/
-// inspiration from https://github.com/ChainSafe/web3-context
-
 import React, { createContext, useEffect, useMemo, useRef } from 'react';
-// import Onboard from '@tracer-protocol/onboard';
 import { API as OnboardApi, Initialization, Wallet } from '@tracer-protocol/onboard/dist/src/interfaces';
 import { Network } from './Web3Context.Config';
 import { ethers, providers } from 'ethers';
-// import { ARBITRUM } from '~/constants/networks';
 import { KnownNetwork } from '@tracer-protocol/pools-js';
 import { useStore } from '@store/main';
 import { selectIsDark } from '@store/ThemeSlice';
-import {selectWeb3Slice} from '@store/Web3Slice';
+import { selectWeb3Slice } from '@store/Web3Slice';
 
 export type OnboardConfig = Partial<Omit<Initialization, 'networkId'>>;
 
@@ -59,14 +54,8 @@ const Web3Store: React.FC<Web3ContextProps> = ({
 }) => {
     const isDark = useStore(selectIsDark);
 
-
-    const { 
-        onboard, account, provider, 
-        network, wallet, isReady, 
-        handleConnect,
-        checkIsReady,
-        resetOnboard
-    } = useStore(selectWeb3Slice);
+    const { onboard, account, provider, network, wallet, isReady, handleConnect, checkIsReady, resetOnboard } =
+        useStore(selectWeb3Slice);
     // const [account, setAccount] = useState<string | undefined>(undefined);
     // const [signer, setSigner] = useState<ethers.Signer | undefined>(undefined);
     // const [network, setNetwork] = useState<KnownNetwork | undefined>(undefined);
@@ -83,101 +72,101 @@ const Web3Store: React.FC<Web3ContextProps> = ({
 
     // Initialize OnboardJS
     // useEffect(() => {
-        // const initializeOnboard = async () => {
-            // const checks = 
-                // [{ checkName: 'accounts' }, { checkName: 'connect' }];
-            // try {
-                // const onboard = Onboard({
-                    // ...onboardConfig,
-                    // networkId: networkIds ? networkIds[0] : parseInt(ARBITRUM), //Default to arb
-                    // darkMode: isDark,
-                    // walletCheck: checks,
-                    // subscriptions: {
-                        // address: (address) => {
-                            // console.info(`Changing address: ${address}`);
-                            // setAccount(address);
-                            // onboardConfig?.subscriptions?.address && onboardConfig?.subscriptions?.address(address);
-                        // },
-                        // wallet: (wallet) => {
-                            // console.debug('Detected wallet change');
-                            // if (wallet.provider) {
-                                // console.debug('Setting wallet provider');
-                                // if (wallet.name && cacheWalletSelection) {
-                                    // window.localStorage.setItem('onboard.selectedWallet', wallet.name);
-                                // }
-                                // const provider_ = new ethers.providers.Web3Provider(wallet.provider, 'any');
-                                // console.debug('Waiting for injected wallet provider');
-                                // provider_.ready.then(() => {
-                                    // console.debug('Injected wallet provider ready');
-                                    // console.debug('Setting injected wallet provider', provider_);
-                                    // setWallet(wallet);
-                                    // usingDefaultProvider.current = false;
-                                    // setProvider(provider_);
-                                    // if (provider_?.network.chainId) {
-                                        // setNetwork(provider_.network.chainId.toString() as KnownNetwork);
-                                    // }
-                                // });
-                            // } else {
-                                // setWallet(undefined);
-                            // }
-                            // onboardConfig?.subscriptions?.wallet && onboardConfig.subscriptions.wallet(wallet);
-                        // },
-                        // network: (network) => {
-                            // if (!networkIds || networkIds.includes(network)) {
-                                // onboard.config({ networkId: network });
-                            // }
-                            // console.info(`Changing network ${network}`);
-                            // const network_ = network?.toString() as KnownNetwork;
-                            // setNetwork(network_);
-                            // setConfig(networkConfig[network_ ?? DEFAULT_NETWORK]);
-                            // onboardConfig?.subscriptions?.network && onboardConfig.subscriptions.network(network);
-                        // },
-                    // },
-                // });
+    // const initializeOnboard = async () => {
+    // const checks =
+    // [{ checkName: 'accounts' }, { checkName: 'connect' }];
+    // try {
+    // const onboard = Onboard({
+    // ...onboardConfig,
+    // networkId: networkIds ? networkIds[0] : parseInt(ARBITRUM), //Default to arb
+    // darkMode: isDark,
+    // walletCheck: checks,
+    // subscriptions: {
+    // address: (address) => {
+    // console.info(`Changing address: ${address}`);
+    // setAccount(address);
+    // onboardConfig?.subscriptions?.address && onboardConfig?.subscriptions?.address(address);
+    // },
+    // wallet: (wallet) => {
+    // console.debug('Detected wallet change');
+    // if (wallet.provider) {
+    // console.debug('Setting wallet provider');
+    // if (wallet.name && cacheWalletSelection) {
+    // window.localStorage.setItem('onboard.selectedWallet', wallet.name);
+    // }
+    // const provider_ = new ethers.providers.Web3Provider(wallet.provider, 'any');
+    // console.debug('Waiting for injected wallet provider');
+    // provider_.ready.then(() => {
+    // console.debug('Injected wallet provider ready');
+    // console.debug('Setting injected wallet provider', provider_);
+    // setWallet(wallet);
+    // usingDefaultProvider.current = false;
+    // setProvider(provider_);
+    // if (provider_?.network.chainId) {
+    // setNetwork(provider_.network.chainId.toString() as KnownNetwork);
+    // }
+    // });
+    // } else {
+    // setWallet(undefined);
+    // }
+    // onboardConfig?.subscriptions?.wallet && onboardConfig.subscriptions.wallet(wallet);
+    // },
+    // network: (network) => {
+    // if (!networkIds || networkIds.includes(network)) {
+    // onboard.config({ networkId: network });
+    // }
+    // console.info(`Changing network ${network}`);
+    // const network_ = network?.toString() as KnownNetwork;
+    // setNetwork(network_);
+    // setConfig(networkConfig[network_ ?? DEFAULT_NETWORK]);
+    // onboardConfig?.subscriptions?.network && onboardConfig.subscriptions.network(network);
+    // },
+    // },
+    // });
 
-                // const savedWallet = window.localStorage.getItem('onboard.selectedWallet');
-                // if (cacheWalletSelection && savedWallet) {
-                    // (async () => {
-                        // await onboard.walletSelect(savedWallet);
-                        // await onboard.walletCheck();
-                        // setOnboard(onboard);
-                    // })();
-                // } else {
-                    // setOnboard(onboard);
-                // }
-            // } catch (error) {
-                // console.error('Error initializing onboard', error);
-            // }
-        // };
+    // const savedWallet = window.localStorage.getItem('onboard.selectedWallet');
+    // if (cacheWalletSelection && savedWallet) {
+    // (async () => {
+    // await onboard.walletSelect(savedWallet);
+    // await onboard.walletCheck();
+    // setOnboard(onboard);
+    // })();
+    // } else {
+    // setOnboard(onboard);
+    // }
+    // } catch (error) {
+    // console.error('Error initializing onboard', error);
+    // }
+    // };
 
-        // initializeOnboard();
+    // initializeOnboard();
     // }, []);
 
     // useEffect(() => {
-        // let mounted = true;
-        // const waitForDefaultProvider = async () => {
-            // if (DEFAULT_WSS_RPC) {
-                // const provider_ = new ethers.providers.WebSocketProvider(DEFAULT_WSS_RPC);
-                // websocket providers need to initiate
-                // console.debug('Waiting for provider', provider_.ready);
-                // await provider_.ready;
-                // console.debug('Provider ready, setting provider', usingDefaultProvider.current);
-                // if (usingDefaultProvider.current) {
-                    // if the provider has not been set by onboard
-                    // if (mounted) {
-                        // setNetwork(DEFAULT_NETWORK);
-                        // setConfig(networkConfig[DEFAULT_NETWORK]);
-                        // setProvider(provider_);
-                    // }
-                // }
-            // }
-        // };
-        // if (onboard && checkIsReady()) {
-            // waitForDefaultProvider();
-        // }
-        // return () => {
-            // mounted = false;
-        // };
+    // let mounted = true;
+    // const waitForDefaultProvider = async () => {
+    // if (DEFAULT_WSS_RPC) {
+    // const provider_ = new ethers.providers.WebSocketProvider(DEFAULT_WSS_RPC);
+    // websocket providers need to initiate
+    // console.debug('Waiting for provider', provider_.ready);
+    // await provider_.ready;
+    // console.debug('Provider ready, setting provider', usingDefaultProvider.current);
+    // if (usingDefaultProvider.current) {
+    // if the provider has not been set by onboard
+    // if (mounted) {
+    // setNetwork(DEFAULT_NETWORK);
+    // setConfig(networkConfig[DEFAULT_NETWORK]);
+    // setProvider(provider_);
+    // }
+    // }
+    // }
+    // };
+    // if (onboard && checkIsReady()) {
+    // waitForDefaultProvider();
+    // }
+    // return () => {
+    // mounted = false;
+    // };
     // }, [onboard]);
 
     useEffect(() => {
@@ -187,18 +176,18 @@ const Web3Store: React.FC<Web3ContextProps> = ({
     }, [isDark]);
 
     // useEffect(() => {
-        // const signer = provider?.getSigner();
-        // setSigner(signer);
+    // const signer = provider?.getSigner();
+    // setSigner(signer);
     // }, [provider, account]);
 
     useMemo(() => {
         let mounted = true;
         if (provider) {
             // provider.getBlockNumber().then((num) => {
-                // console.debug(`Setting block number: ${num}`);
-                // if (mounted) {
-                    // setBlockNumber(num);
-                // }
+            // console.debug(`Setting block number: ${num}`);
+            // if (mounted) {
+            // setBlockNumber(num);
+            // }
             // });
             provider.getGasPrice().then((gasPrice) => {
                 console.debug(`Setting gas price: ${gasPrice.toNumber()}`);
@@ -213,29 +202,29 @@ const Web3Store: React.FC<Web3ContextProps> = ({
     }, [provider, network]);
 
     // const checkIsReady = async () => {
-        // const isReady = await onboard?.walletCheck().catch((_err) => false);
-        // console.debug('Wallet is ready', isReady);
-        // setIsReady(!!isReady);
-        // return !!isReady;
+    // const isReady = await onboard?.walletCheck().catch((_err) => false);
+    // console.debug('Wallet is ready', isReady);
+    // setIsReady(!!isReady);
+    // return !!isReady;
     // };
 
     // const resetOnboard = async () => {
-        // window.localStorage.setItem('onboard.selectedWallet', '');
-        // setIsReady(false);
-        // await onboard?.walletReset();
+    // window.localStorage.setItem('onboard.selectedWallet', '');
+    // setIsReady(false);
+    // await onboard?.walletReset();
     // };
 
     // const handleConnect = async () => {
-        // if (onboard) {
-            // try {
-                // const selectedWallet = await onboard?.walletSelect();
-                // if (selectedWallet) {
-                    // await checkIsReady();
-                // }
-            // } catch (err) {
-                // console.error(err);
-            // }
-        // }
+    // if (onboard) {
+    // try {
+    // const selectedWallet = await onboard?.walletSelect();
+    // if (selectedWallet) {
+    // await checkIsReady();
+    // }
+    // } catch (err) {
+    // console.error(err);
+    // }
+    // }
     // };
 
     const web3Context = React.useMemo(
