+++ hooks/useBalancerSpotPrices/index.tsx
@@ -2,14 +2,15 @@ import { useMemo, useState } from 'react';
 import BigNumber from 'bignumber.js';
 import { KnownNetwork } from '@tracer-protocol/pools-js';
 import { getBalancerPrices } from '~/utils/balancer';
-import { DEFAULT_NETWORK, networkConfig } from '@context/Web3Context/Web3Context.Config';
+import { DEFAULT_NETWORK } from '@context/Web3Context/Web3Context.Config';
+import { balancerConfig } from '~/constants/balancer';
 
 export default ((network = DEFAULT_NETWORK) => {
     const [tokenPrices, setTokenPrices] = useState<Record<string, BigNumber>>({});
 
     useMemo(() => {
         let mounted = true;
-        getBalancerPrices(networkConfig[network]?.balancerInfo).then((tokenPrices) => {
+        getBalancerPrices(balancerConfig[network]).then((tokenPrices) => {
             if (mounted) {
                 setTokenPrices(tokenPrices);
             }
