+++ pages/_app.tsx
@@ -10,9 +10,7 @@ import '../styles/index.css';
 import { ToastContainerWithStyles } from '@components/General/Notification/ToastContainerWithStyles';
 import { Web3Store } from '@context/Web3Context/Web3Context';
 import { UsersCommitStore } from '@context/UsersCommitContext';
-import { networkConfig } from '@context/Web3Context/Web3Context.Config';
 import { StyledThemeProvider } from '@context/ThemeContext';
-import { ARBITRUM, ARBITRUM_RINKEBY, MAINNET } from '~/constants/networks';
 
 const USERSNAP_GLOBAL_API_KEY = process.env.NEXT_PUBLIC_USERSNAP_GLOBAL_API_KEY;
 const USERSNAP_API_KEY = process.env.NEXT_PUBLIC_USERSNAP_API_KEY;
@@ -59,33 +57,7 @@ const App = ({ Component, pageProps }: AppProps) => { // eslint-disable-line
                 <script defer data-domain="pools.tracer.finance" src="https://plausible.io/js/plausible.js" />
             </Head>
             <StyledThemeProvider>
-                <Web3Store
-                    onboardConfig={{
-                        hideBranding: true,
-                        walletSelect: {
-                            heading: 'Connect Wallet',
-                            wallets: [
-                                { walletName: 'metamask' },
-                                { walletName: 'coinbase' },
-                                { walletName: 'torus' },
-                                // { walletName: "binance" },
-
-                                {
-                                    walletName: 'walletConnect',
-                                    rpc: {
-                                        [ARBITRUM]: networkConfig[ARBITRUM].publicRPC,
-                                        [ARBITRUM_RINKEBY]: networkConfig[ARBITRUM_RINKEBY].publicRPC,
-                                        [MAINNET]: networkConfig[MAINNET].publicRPC,
-                                    },
-                                },
-                            ],
-                            // agreement: {
-                            //     version: '1.0',
-                            //     termsUrl: 'https://google.com',
-                            // },
-                        },
-                    }}
-                >
+                <Web3Store>
                     <UsersCommitStore>
                         <Component {...pageProps} />
                     </UsersCommitStore>
