+++ store/GasSlice/types.ts
@@ -1,5 +1,5 @@
 export interface IGasSlice {
-    gasPrice: number,
-    fetchingGasPrice: boolean,
-    getGasPrice: () => void
+    gasPrice: number;
+    fetchingGasPrice: boolean;
+    getGasPrice: () => void;
 }
