+++ store/Web3Slice/types.ts
@@ -1,5 +1,5 @@
-import { ethers } from "ethers";
-import { KnownNetwork } from "@tracer-protocol/pools-js";
+import { ethers } from 'ethers';
+import { KnownNetwork } from '@tracer-protocol/pools-js';
 import { API as OnboardApi, Wallet } from '@tracer-protocol/onboard/dist/src/interfaces';
 
 export interface IWeb3Slice {
@@ -10,10 +10,9 @@ export interface IWeb3Slice {
     provider: ethers.providers.JsonRpcProvider | undefined;
     wallet: Wallet | undefined;
 
-    checkIsReady: () => Promise<boolean>
-    resetOnboard: () => Promise<void>
-    handleConnect: () => Promise<void>
-
+    checkIsReady: () => Promise<boolean>;
+    resetOnboard: () => Promise<void>;
+    handleConnect: () => Promise<void>;
 
     // unsupportedNetworkPopupRef: React.MutableRefObject<string>;
 }
