+++ types/balancer.ts
@@ -1,0 +1,7 @@
+export type BalancerInfo = {
+    graphUri: string;
+    baseUri: string; // base link to balancer trading page
+    pools: string[];
+    leveragedPools: string[];
+    wPool: string;
+};
