+++ types/networks.ts
@@ -1,0 +1,3 @@
+import { UNKNOWN_NETWORK } from '~/constants/networks';
+
+export type UnknownNetwork = typeof UNKNOWN_NETWORK;
