+++ utils/rpcMethods.ts
@@ -1,7 +1,7 @@
-import { tokenSymbolToLogoTicker } from '@components/General';
-import { networkConfig } from '@context/Web3Context/Web3Context.Config';
 import { ethers } from 'ethers';
 import { KnownNetwork } from '@tracer-protocol/pools-js';
+import { tokenSymbolToLogoTicker } from '@components/General';
+import { networkConfig } from '@context/Web3Context/Web3Context.Config';
 
 const tokenImagesRootUrl = 'http://ipfs.io/ipfs/QmaKrQSyTSdcmikLHdtKHp6tn3pT3Gcnu2BWziN97Fscrd';
 /**
@@ -113,7 +113,7 @@ export const openArbiscan: (type: ArbiscanEnum, taraget: string, network: KnownN
 
 export function isAddress(value: string): boolean {
     try {
-        return !!ethers.utils.getAddress(value)
+        return !!ethers.utils.getAddress(value);
     } catch {
         return false;
     }
